[chore] migrate of functions
